// modules/apps/headless/headless-admin-site/headless-admin-site-client-js/src/models/HtmlProperties.ts
/**
 * SPDX-FileCopyrightText: (c) 2026 Liferay, Inc. https://liferay.com
 * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
 */


/**
 * @author Rubén Pulido
 * @generated
 */

	export class HtmlProperties {
			"htmlTag"?: 'Article' | 'Aside' | 'Div' | 'Footer' | 'Header' | 'Nav' | 'Section';

		static "discriminator": string | undefined = undefined;

	static "attributeTypeMap": Array<{
		baseName: string;
		name: string;
		type: string;
	}> = [
		{
			baseName: "htmlTag",
			name: "htmlTag",
			type: "'Article' | 'Aside' | 'Div' | 'Footer' | 'Header' | 'Nav' | 'Section'",
		},
		];

		static getAttributeTypeMap() {
				return HtmlProperties.attributeTypeMap;
		}
	}